requirements.txt
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
Faker==15.3.4
ipython==8.7.0
jedi==0.18.2
matplotlib-inline==0.1.6
mysqlclient==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
toposort==1.7
traitlets==5.7.1
wcwidth==0.2.5
